backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pgvector

pydantic==2.12.5
pydantic-settings==2.10.1

bcrypt
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

redis==5.0.4

python-dotenv==1.2.1
email-validator==2.2.0
python-multipart==0.0.20
httpx==0.26.0
resend==2.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
